Implement remainder of #SPC-sel4syscrate.cmake
This required adding code to define specific CMake variables depending
on the cargo profile (debug or release) we are building for. One
addtional CMake variable that this depends on is KernelVerificationBuild
(which was added to the list in #SPC-sel4syscrate).

Added a target_spec file to allow checking the implementation on an
Aarch32 target. Made the sel4-sys crate itself #[no_std] to keep
everything compline (under xargo anyways).