Implement SPC-sel4platcrate.buildcrate
Extracted the cmake and bindgen functionality of the sel4-sys build.rs
file into the sel4-build crate. The root CMakeLists.txt file and the
flags.cmake file that it includes were also moved to the sel4-build
crate.

The new CMakeLists.txt file in the sel4-build crate now has to be a
template into which the manifest directory is inserted. The purpose of
inserting the manifset directlty is to locate the seL4 source code
(which is presumed to be in the seL4 subdirectory of the manifest
directory).

The uses of the askama template engine to drive the rendering of the
CMakeLists.txt file may possibly be overkill, but it gets the job done.