Make sel4-build a no-op when target_os = "none"
This change permits a no-op build for sel4-build for the
armv7-unknown-none-eabihf and x86_64-unknown-none-elf target triples
which simplifies running cargo subcommands on all subprojects of the
virtual manifest in the workspace root.

sel4-build is built for the host triple (which is not target_os =
"none") when it is being used for as part of a build script. Also, the
sel4-sys tests that use it are also conditional on not(target_os =
"none") so they will be run in a hosted (i.e. not no_std) build of
seL4-sys.