Remove the sel4-plat-kzm crate
The kzm platform crate was not properly building the seL4 microkernal
and its page on the "Supported Hardware" section of the seL4 website
indicates that "The KZM is depreciated". Removing support for a
depreciated platform was easier than trying to figure out why the
relevement CMake invocation (which in turn was invocing Ninja) was
ending in a build error.

This commit also removes SPC-sel4platcrate.kzm and adds an explination
for its absense to SPC-sel4platcrate.