Fix binding generation for platforms
The bindings that were being generated for the sel4-plat-* crates (with
CMakeTarget::Kernel(platform)) were  not including platform specifc
constants that are dependant on HARDWARE_DEBUG_API being set. The
assumption for building both the kernel and libsel4 is that
HARDWARE_DEBUG_API is set for the Debug profile and not for the Release
profile.

This commit explitly defines CONFIG_HARDWARE_DEBUG_API for the binding
generation for Kernel(platform) builds when the profile is Debug. This
will generate the required platform specific, HARDWARE_DEBUG_API
dependant constants.