Modify REQ-purpose to list supported architectures
The change lists x86_64, aarch32, and aarch64 as the initially supported
architectures. It explicitly leaves off ia32 (aka i686) and implicitly
leaves off all risc-v platforms.

We leave off risc-v platforms because their support in both seL4 and in
rustc is still experimental. Once support in both of these places mature
it may be possible to revisit the inclusion of risc-v platforms.

I attempted to formulate a target spec json file for ia32 to be able to
support that file, but was not able to get "cargo xbuild" to be able
compile the sysroot using that target. I was also unable to locate an
example target spec file for ia32 for any other project. It is also not
clear how important ia32 support is in practice.