Remove description of ia32 support for pc99 platform
The project as a whole is not supporting ia32 so this should not be
listed in SPC-sel4platcrate.pc99.

There is still code in sel4-build to support ia32 but in the absence of
a target spec json file there hasn't been a test that this code builds
libsel4 or kernel.elf as expected so the design artifact shouldn't
suggest that this has been implemented.