Implement SPC-sel4platcrate.hikey
This is the first AArch64 platform implemented. The attempt to implement
it with the current assumpation about the HardwareDebugAPI configuration
option showed that that option is not supported in seL4 (as of 10.1.1)
for any AArch64 platforms. This required changing the assumptions in
SPC-sel4syscrate about this configuration option and adapting the
sel4-build crate to account for these changes.

The HardwareDebugAPI configuration option is now on for the Debug
profile for all architectures except AArch64. It remains off for the
Release profile on all architectures.