Change term for non-ABI seL4 Kernel* options
This changes the name used for the groupings of non-ABI seL4 Kernel*
options from "profiles" to "classes". The term "profiles" is already
meaningful for cargo and we already used it to work differently for
debug and release builds. Overloading "profile" could be confusing in
this context.

Although "class" is used in other contexts it is not broadly used in
Rust and also does not appear to be used in seL4. This should minimize
the risk of confusion.