Add initial Travis CI build file
This .travis.yml file is a minimal one that likely will not build
because it is missing some required dependancies. A build without these,
though will help identifiy what is missing.

The eventual CI build will also have a matrix to cross complie for the
various "none" targets as well as for the host.