Change sel4-sys to not link to libsel4.a on hosted build
This also required changing the sel4-sys smoketest to compile (for real)
only on the "none" os platforms. This exposed an inconsistancy in the
handline of floating point for the armv4-unknown-non-eabihf target. This
was fixed by changing the target to armv7-unknown-none-eabi and changing
the target spec file to (try to) exclude all hard floating point.