Add Travis CI dependancy on libc6-dev-armel-cross
This is a recommend dependancy of the gcc-arm-linux-gnueabi package but
Travis CI does not automaticlly install recommend dependancies in an apt
addon step. This package is needed for some of the steps compiling seL4.

In particular, the tests to see if the -mfloat-abi=soft flag is
supported try to compile an arm linux hosted binary (a simple
"void main(void) {}" C program) with that flag. This was resulting in a
link error on Travis because ln could not find crt1.o and crti.o (which
are in libc6-dev-armel-cross).

This was a "Works on my machine" build failure because the default for
apt-get is to install recommended packages so an "apt-get
gcc-arm-linux-gnueabi" on "my machine" would install the recommended
"libc6-dev-armel-cross" package while Travis CI would not install this
recommended package.