Implement Layout and StartPaddr::Specified layout strategy
Removed Input::output_size() because it is not possible to determint the
size of the output independent of the layout strategy used and there is
already a required_size() method on Layout.

Added Arch type to carry the "architecture" information from the ELF
header from the Input, through the Layout, into the file written by the
OutputWriter. This "architecture" information, the file's entry point
together with the output program header offset is enough to create the
ELF header for the output file.