Implement elf-preload OutputWriter
OutputWriter is implemented by delegating to a contained ElfWriter which
co-ordinates writing the segments, but calls back into Layout for the
actual writing.

The call backs into Layout are abstracted through the SegmentWriter
trait which is local to the output module. The Layout methods to which
the SegmentWriter implementation calls are currently unimplemented.