Implement Layout::segment_size() and write_segment()
The implementation required changes to Input to extract the entry point
from the Elf header of the input file and then carry it through the
layout so it is available in write_segment() when writing the segment
that includes the Elf header in the output file.

Changes to LayoutStrategy to localize the assumptions about the mapping
of loadable segment number to an index in the input and output
ProgramHeader slices.

Incidental changes to OutputWriter to properly pass errors from
write_segment() back out.