Change one test criterion for elf-preload
The #TST-elfpreload.idempotent test design had called for using
"objcopy -O binary ..." to check for idempotentcy. This was supposed to
be a very clever way of checking that the file could be memcpy'd to its
physical load address as one operation (instead of segment by segment).

Unfortunatly this is too clever. "objcopy" works on the sections (not
segments) of its input file so it fails to read the output "elf-preload"
which does not have any section headers. The lack of section headers is
a requirement of #SPC-elfpreload.programheader.

The revised test criterion for #TST-elfpreload.idempotent relies on
walking the PT_LOAD program headers of the output of "elf-preload" to
check directly that they account for every byte of the output.