Start transition to cargo make
This commit adds the basic structure of the Makefile.toml files with an
"xbuild" task that is conditional on XBUILD_TARGET being set and making
the default "build" task conditional on XBUILD_TARGET not being set.

The Makefile.toml in each member of the workspace just extends the root
Makefile.toml for now.